Refactoring. Created error and success modules.
